fastapi_app/requirements.txt
# FastAPI Core
fastapi==0.110.2
uvicorn==0.29.0
pydantic==2.7.0
pydantic_core==2.18.1

# Database
mysql-connector-python==8.4.0

# Cache
redis==5.0.7

# Async Support
aiohttp==3.9.5

# Security & Authentication
bcrypt==4.1.3
PyJWT==2.8.0
python-dotenv==1.0.1

# Utils
shortuuid==1.0.13

